refactor: 예약 취소 책임 분리
