+++ src/test/java/roomescape/reservation/ReservationServiceTest.java
@@ -16,6 +16,8 @@ import roomescape.domain.theme.Theme;
 import roomescape.exception.ConflictException;
 import roomescape.exception.InvalidInputException;
 import roomescape.exception.ResourceNotFoundException;
+import roomescape.service.reservation.ReservationCancellationService;
+import roomescape.service.reservation.ReservationFactory;
 import roomescape.service.reservation.ReservationService;
 import roomescape.service.reservationtime.ReservationTimeService;
 import roomescape.service.theme.ThemeService;
@@ -222,6 +224,12 @@ class ReservationServiceTest {
         private final FakeReservationSlotRepository reservationSlotRepository = new FakeReservationSlotRepository();
         private final FakeReservationWaitingRepository reservationWaitingRepository = new FakeReservationWaitingRepository();
         private final FakeReservationTimeRepository reservationTimeRepository = new FakeReservationTimeRepository();
+        private final ReservationFactory reservationFactory = new ReservationFactory();
+        private final ReservationCancellationService reservationCancellationService = new ReservationCancellationService(
+                reservationRepository,
+                reservationWaitingRepository,
+                reservationFactory
+        );
         private final ThemeService themeService = new ThemeService(themeRepository, reservationRepository);
         private final ReservationTimeService reservationTimeService = new ReservationTimeService(
                 reservationTimeRepository,
@@ -232,7 +240,8 @@ class ReservationServiceTest {
         private final ReservationService reservationService = new ReservationService(
                 reservationRepository,
                 reservationSlotRepository,
-                reservationWaitingRepository,
+                reservationCancellationService,
+                reservationFactory,
                 reservationTimeService,
                 themeService
         );
